修改了一些类的代码，并且重新定义了价格类

+++ src/main/java/com/example/nftmarket/entity/Dinosaur.java
@@ -16,7 +16,6 @@ public class Dinosaur {
     Boolean DinosaurSex;
     String DinosaurType;
     Boolean isBreeding;
-
     //Details
     String DinosaurColor;
     String DinosaurClothing;
@@ -24,5 +23,5 @@ public class Dinosaur {
     String DinosaurOwner;
     String DinosaurRarity;
     String DinosaurPhotoUri;
-    Integer DinosaurPrice;
+    Double DinosaurPrice;
 }
